+++ packages/example/src/adapters/user_repository_adapter.rs
@@ -1,29 +1,46 @@
 use std::sync::Arc;
 
 use async_trait::async_trait;
+use ioc_container_rs::{
+  errors::error::Error,
+  ports::{adapter_port::AdapterPort, context_port::ContextPort},
+};
 use tokio::sync::RwLock;
 
 use crate::{
   entities::user_entity::UserEntity, ports::output::user_repository_port::UserRepositoryPort,
 };
 
-pub struct Adapter {
+pub struct UserRepositoryAdapter {
   store: Arc<RwLock<Vec<UserEntity>>>,
 }
 
-impl Adapter {
+impl UserRepositoryAdapter {
   pub fn new(store: Arc<RwLock<Vec<UserEntity>>>) -> Self {
     Self { store }
-  }
-
-  pub fn token() -> &'static str {
-    "USER_REPOSITORY_ADAPTER"
   }
 }
 
 #[async_trait]
-impl UserRepositoryPort for Adapter {
-  async fn add_user(&self, entity: &UserEntity) -> Result<(), String> {
+impl AdapterPort<UserRepositoryAdapter> for UserRepositoryAdapter {
+  fn token() -> &'static str {
+    "USER_REPOSITORY_ADAPTER"
+  }
+
+  async fn get_adapter(context: &Arc<dyn ContextPort>) -> Result<Box<Self>, Error> {
+    let me = context
+      .resolve_provider(Self::token())
+      .await?
+      .downcast::<Self>()
+      .map_err(|_| format!("Cant resolve provider: {}", Self::token()))?;
+
+    Ok(me)
+  }
+}
+
+#[async_trait]
+impl UserRepositoryPort for UserRepositoryAdapter {
+  async fn add_user(&self, entity: &UserEntity) -> Result<(), Error> {
     let mut store = self.store.write().await;
 
     store.push(entity.clone());
@@ -31,7 +48,7 @@ impl UserRepositoryPort for Adapter {
     Ok(())
   }
 
-  async fn delete_user(&self, entity: &UserEntity) -> Result<(), String> {
+  async fn delete_user(&self, entity: &UserEntity) -> Result<(), Error> {
     let mut store = self.store.write().await;
 
     let index = store.iter().position(|u| u.email == entity.email);
@@ -41,11 +58,11 @@ impl UserRepositoryPort for Adapter {
         store.remove(i);
         Ok(())
       }
-      None => Err("User not found for delete.".to_string()),
+      None => Err("User not found for delete.".into()),
     }
   }
 
-  async fn get_user_by_email(&self, email: &str) -> Result<Option<UserEntity>, String> {
+  async fn get_user_by_email(&self, email: &str) -> Result<Option<UserEntity>, Error> {
     let store = self.store.read().await;
 
     let index = store.iter().position(|u| u.email == email);
@@ -59,21 +76,7 @@ impl UserRepositoryPort for Adapter {
     }
   }
 
-  async fn get_user(&self, entity: &UserEntity) -> Result<Option<UserEntity>, String> {
-    let store = self.store.read().await;
-
-    let index = store.iter().position(|u| u.email == entity.email);
-
-    match index {
-      Some(i) => match store.get(i) {
-        Some(user) => Ok(Some(user.clone())),
-        _ => Ok(None),
-      },
-      _ => Ok(None),
-    }
-  }
-
-  async fn get_count(&self) -> Result<usize, String> {
+  async fn get_count(&self) -> Result<usize, Error> {
     let store = self.store.read().await;
 
     Ok(store.len())