+++ packages/example/src/lib.rs
@@ -1,56 +1,24 @@
-use std::sync::Arc;
-
-use adapters::{user_repository_adapter, user_service_adapter};
-use entities::user_entity::UserEntity;
-use ioc_container_rs::{
-  container::di::{InjectAdapter, DI},
-  context::{container_context::ContainerContext, context::Context},
-};
-use tokio::sync::RwLock;
-
-mod adapters;
-mod entities;
-mod ports;
-
-pub async fn create_di() -> DI {
-  let store: Arc<RwLock<Vec<UserEntity>>> = Arc::new(RwLock::new(vec![]));
-
-  let user_repository_injector = InjectAdapter {
-    token: user_repository_adapter::Adapter::token(),
-    factory: Arc::new(move |_| user_repository_adapter::Adapter::new(store.clone())),
-  };
-
-  let di = DI::new()
-    .inject(user_repository_injector)
-    .await
-    .inject(InjectAdapter {
-      token: user_service_adapter::Adapter::token(),
-      factory: Arc::new(user_service_adapter::Adapter::new),
-    })
-    .await;
-
-  di
-}
-
-pub async fn get_user_service(context: &Arc<ContainerContext>) -> user_service_adapter::Adapter {
-  context
-    .resolve_provider::<user_service_adapter::Adapter>(user_service_adapter::Adapter::token())
-    .await
-}
-
-pub async fn get_user_repository(
-  context: &Arc<ContainerContext>,
-) -> user_repository_adapter::Adapter {
-  context
-    .resolve_provider::<user_repository_adapter::Adapter>(user_repository_adapter::Adapter::token())
-    .await
-}
+pub mod adapters;
+pub mod entities;
+pub mod ports;
 
 #[cfg(test)]
 mod tests {
 
+  use std::sync::Arc;
+
+  use ioc_container_rs::{
+    container::di::{InjectAdapter, DI},
+    context::container_context::ContainerContext,
+    ports::adapter_port::AdapterPort,
+  };
+  use tokio::sync::RwLock;
+
   use crate::{
-    create_di, get_user_repository, get_user_service,
+    adapters::{
+      user_repository_adapter::UserRepositoryAdapter, user_service_adapter::UserServiceAdapter,
+    },
+    entities::user_entity::UserEntity,
     ports::{
       input::{
         add_user_port::AddUserPort, delete_user_port::DeleteUserPort, get_user_port::GetUserPort,
@@ -59,17 +27,50 @@ mod tests {
     },
   };
 
+  async fn create_di() -> DI {
+    let di = DI::new(Arc::new(ContainerContext::new()));
+
+    let store: Arc<RwLock<Vec<UserEntity>>> = Arc::new(RwLock::new(vec![]));
+
+    let user_repository_injector = InjectAdapter {
+      token: UserRepositoryAdapter::token(),
+      factory: Arc::new(move |_| UserRepositoryAdapter::new(store.clone())),
+    };
+
+    let di = di
+      .inject(user_repository_injector)
+      .await
+      .expect("UserRepositoryAdapter should be injected");
+
+    let di = di
+      .inject(InjectAdapter {
+        token: UserServiceAdapter::token(),
+        factory: Arc::new(UserServiceAdapter::new),
+      })
+      .await
+      .expect("UserServiceAdapter should be injected");
+
+    di
+  }
+
   #[tokio::test]
   async fn should_be_return_zero_count_from_repository() {
     let di = create_di().await;
 
     let context = di.get_context();
 
-    let user_repository = get_user_repository(&context).await;
+    let user_repository = UserRepositoryAdapter::get_adapter(&context).await;
 
-    let count = user_repository.get_count().await.unwrap();
+    assert!(user_repository.is_ok());
 
-    assert_eq!(count, 0);
+    let user_repository = user_repository.expect("UserRepository should exist");
+
+    let count = user_repository
+      .get_count()
+      .await
+      .expect("Get count has error");
+
+    assert_eq!(count, 0, "Count should be 0");
   }
 
   #[tokio::test]
@@ -78,11 +79,13 @@ mod tests {
 
     let context = di.get_context();
 
-    let user_service = get_user_service(&context).await;
+    let user_service = UserServiceAdapter::get_adapter(&context)
+      .await
+      .expect("UserService should exist");
 
-    let count = user_service.get_count().await.unwrap();
+    let count = user_service.get_count().await.expect("Get count has error");
 
-    assert_eq!(count, 0);
+    assert_eq!(count, 0, "Count should be 0");
   }
 
   #[tokio::test]
@@ -91,7 +94,9 @@ mod tests {
 
     let context = di.get_context();
 
-    let user_service = get_user_service(&context).await;
+    let user_service = UserServiceAdapter::get_adapter(&context)
+      .await
+      .expect("UserService should exist");
 
     let user_entity = user_service
       .add_user(&AddUserPort {
@@ -99,12 +104,12 @@ mod tests {
         email: "andrey@mail.ru".to_string(),
       })
       .await
-      .unwrap();
+      .expect("Add user has error");
 
-    let count = user_service.get_count().await.unwrap();
+    let count = user_service.get_count().await.expect("Get count has error");
 
-    assert_eq!(count, 1);
-    assert_eq!(user_entity.name, "Andrey");
+    assert_eq!(count, 1, "Count should be 1");
+    assert_eq!(user_entity.name, "Andrey", "Name should be Andrey");
   }
 
   #[tokio::test]
@@ -113,7 +118,9 @@ mod tests {
 
     let context = di.get_context();
 
-    let user_service = get_user_service(&context).await;
+    let user_service = UserServiceAdapter::get_adapter(&context)
+      .await
+      .expect("UserService should exist");
 
     let user_entity = user_service
       .add_user(&AddUserPort {
@@ -121,23 +128,25 @@ mod tests {
         email: "andrey@mail.ru".to_string(),
       })
       .await
-      .unwrap();
+      .expect("Add user has error");
 
-    let user_deletion = user_service
+    user_service
       .delete_user(&DeleteUserPort {
         email: user_entity.email.to_string(),
       })
-      .await;
-
-    assert_eq!(user_deletion.is_ok(), true);
+      .await
+      .expect("Delete user has error");
 
     let check_user_entity = user_service
       .get_user(&GetUserPort {
         email: user_entity.email.to_string(),
       })
       .await
-      .unwrap();
+      .expect("Get user has error");
 
-    assert_eq!(check_user_entity.is_none(), true);
+    assert!(
+      check_user_entity.is_none(),
+      "User after deletion should be None"
+    );
   }
 }