+++ packages/ioc_container_rs/src/container/di.rs
@@ -1,37 +1,38 @@
 use std::sync::Arc;
 
-use crate::context::{container_context::ContainerContext, context::Context};
+use crate::{
+  errors::error::Error,
+  ports::{adapter_port::AdapterPort, context_port::ContextPort},
+};
 
 pub struct InjectAdapter<T> {
   pub token: &'static str,
-  pub factory: Arc<dyn Fn(Arc<ContainerContext>) -> T + Send + Sync + 'static>,
+  pub factory: Arc<dyn Fn(Arc<dyn ContextPort>) -> T + Send + Sync + 'static>,
 }
 
 #[derive(Clone)]
 pub struct DI {
-  context: Arc<ContainerContext>,
+  context: Arc<dyn ContextPort>,
 }
 
 impl DI {
-  pub fn new() -> Self {
-    let context = Arc::new(ContainerContext::new());
-
+  pub fn new(context: Arc<dyn ContextPort>) -> Self {
     Self { context }
   }
 
-  pub async fn inject<T: Send + Sync + 'static>(&self, injector: InjectAdapter<T>) -> Self {
+  pub async fn inject<T: AdapterPort<T>>(&self, injector: InjectAdapter<T>) -> Result<Self, Error> {
     let container = self.context.get_container();
     let factory = injector.factory;
     let context = self.context.clone();
 
     container
       .register(injector.token, move || factory(context.clone()))
-      .await;
+      .await?;
 
-    self.clone()
+    Ok(self.clone())
   }
 
-  pub fn get_context(&self) -> Arc<ContainerContext> {
+  pub fn get_context(&self) -> Arc<dyn ContextPort> {
     self.context.clone()
   }
 }