+++ packages/ioc_container_rs/src/container/mod.rs
@@ -1,3 +1,2 @@
 pub mod container;
-pub mod container_store;
 pub mod di;