+++ packages/tests/src/adapters/adapter_string_test.rs
@@ -1,3 +1,16 @@
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use ioc_container_rs::{
+  errors::error::Error,
+  ports::{adapter_port::AdapterPort, context_port::ContextPort},
+};
+
+pub trait AdapterStringTestPort {
+  fn get_message(&self) -> &str;
+  fn set_message(&mut self, message: String);
+}
+
 pub struct AdapterStringTest {
   message: String,
 }
@@ -5,19 +18,34 @@ pub struct AdapterStringTest {
 impl AdapterStringTest {
   pub fn new() -> Self {
     Self {
-      message: "Hello, World!".to_string(),
+      message: String::from("Hello, World!"),
     }
   }
+}
 
-  pub fn token() -> &'static str {
+#[async_trait]
+impl AdapterPort<AdapterStringTest> for AdapterStringTest {
+  fn token() -> &'static str {
     "ADAPTER_STRING_TEST"
   }
 
-  pub fn get_message(&self) -> &str {
+  async fn get_adapter(context: &Arc<dyn ContextPort>) -> Result<Box<Self>, Error> {
+    let me = context
+      .resolve_provider(Self::token())
+      .await?
+      .downcast::<Self>()
+      .map_err(|_| format!("Cant resolve provider: {}", Self::token()))?;
+
+    Ok(me)
+  }
+}
+
+impl AdapterStringTestPort for AdapterStringTest {
+  fn get_message(&self) -> &str {
     &self.message
   }
 
-  pub fn set_message(&mut self, message: String) {
+  fn set_message(&mut self, message: String) {
     self.message = message;
   }
 }