+++ packages/tests/src/container/container_single_test.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod tests {
-  use ioc_container_rs::container::container::Container;
+  use ioc_container_rs::{container::container::Container, ports::adapter_port::AdapterPort};
 
   use crate::adapters::adapter_string_test::AdapterStringTest;
 
@@ -8,14 +8,14 @@ mod tests {
   async fn should_be_able_to_register_and_resolve() {
     let container = Container::new();
 
-    container
+    let registered = container
       .register(AdapterStringTest::token(), || AdapterStringTest::new())
       .await;
 
-    let svc = container
-      .resolve::<AdapterStringTest>(AdapterStringTest::token())
-      .await;
+    assert!(registered.is_ok());
 
-    assert_eq!(svc.is_some(), true);
+    let svc = container.resolve(AdapterStringTest::token()).await;
+
+    assert!(svc.is_ok());
   }
 }