+++ packages/tests/src/container_store/mod.rs
@@ -1,3 +1,0 @@
-mod container_store_abstract_test;
-mod container_store_multiple_test;
-mod container_store_single_test;