+++ packages/tests/src/di/di_single_test.rs
@@ -2,50 +2,61 @@
 mod tests {
   use std::sync::Arc;
 
-  use ioc_container_rs::container::di::{InjectAdapter, DI};
-  use ioc_container_rs::context::context::Context;
+  use ioc_container_rs::{
+    container::di::{InjectAdapter, DI},
+    context::container_context::ContainerContext,
+    ports::adapter_port::AdapterPort,
+  };
 
-  use crate::adapters::adapter_string_test::AdapterStringTest;
+  use crate::adapters::adapter_string_test::{AdapterStringTest, AdapterStringTestPort};
+
+  fn create_di() -> DI {
+    DI::new(Arc::new(ContainerContext::new()))
+  }
 
   #[tokio::test]
   async fn should_be_build_and_resolve_service() {
-    let di = DI::new()
+    let di = create_di();
+
+    let di = di
       .inject(InjectAdapter {
         token: AdapterStringTest::token(),
         factory: Arc::new(|_| AdapterStringTest::new()),
       })
-      .await;
+      .await
+      .expect("Failed to inject adapter");
 
     let context = di.get_context();
 
-    context
-      .resolve_provider::<AdapterStringTest>(AdapterStringTest::token())
-      .await;
+    let svc = AdapterStringTest::get_adapter(&context).await;
 
-    assert!(true);
+    assert!(svc.is_ok());
   }
 
   #[tokio::test]
   async fn should_be_mutate_string() {
-    let di = DI::new()
+    let di = create_di();
+
+    let di = di
       .inject(InjectAdapter {
         token: AdapterStringTest::token(),
         factory: Arc::new(|_| AdapterStringTest::new()),
       })
-      .await;
+      .await
+      .expect("Failed to inject adapter");
 
     let context = di.get_context();
 
     const NEW_STRING: &str = "Hello, Rust!";
 
-    let mut svc = context
-      .resolve_provider::<AdapterStringTest>(AdapterStringTest::token())
-      .await;
+    let mut svc = AdapterStringTest::get_adapter(&context)
+      .await
+      .expect("Failed to get adapter");
 
     svc.set_message(NEW_STRING.to_string());
 
     let message = svc.get_message();
 
-    assert_eq!(message, NEW_STRING);
+    assert_eq!(message, NEW_STRING, "Message should be equal");
   }
 }