+++ packages/tests/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod adapters;
 mod container;
-mod container_store;
+pub mod context;
 mod di;